requirements.txt
keyboard==0.13.5
pyautogui==0.9.54
python-socketio==5.9.0
requests==2.31.0
websocket-client==1.6.3
win11toast==0.32
pywebview==4.3.3
python_dotenv==1.0.0